.history/requirements_20240508172828.txt
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
plotly==5.14.0
PyYAML==6.0.1
streamlit==1.28.1
streamlit_authenticator==0.3.1
voici
